# docs/swagger.yaml
definitions:
  wallet.CreateWallet:
    properties:
      balance:
        type: number
      user_id:
        type: integer
      user_name:
        type: string
      wallet_name:
        type: string
      wallet_type:
        type: string
    type: object
  wallet.Err:
    properties:
      message:
        type: string
    type: object
  wallet.UpdateWallet:
    properties:
      balance:
        type: number
      id:
        type: integer
      user_id:
        type: integer
      user_name:
        type: string
      wallet_name:
        type: string
      wallet_type:
        type: string
    type: object
  wallet.Wallet:
    properties:
      balance:
        example: 100
        type: number
      created_at:
        example: "2024-03-25T14:19:00.729237Z"
        type: string
      id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
      user_name:
        example: John Doe
        type: string
      wallet_name:
        example: John's Wallet
        type: string
      wallet_type:
        example: Create Card
        type: string
    type: object
host: localhost:1323
info:
  contact: {}
  description: Sophisticated Wallet API
  title: Wallet API
  version: "1.0"
paths:
  /api/v1/users/{id}/wallets:
    delete:
      consumes:
      - application/json
      description: Delete wallet by user Id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Delete wallet by user Id
      tags:
      - wallet
    get:
      consumes:
      - application/json
      description: Get wallet by user Id
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get wallet by user Id
      tags:
      - wallet
  /api/v1/wallets:
    get:
      consumes:
      - application/json
      description: Get all wallets
      parameters:
      - description: wallet type
        in: query
        name: wallet_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Get all wallets
      tags:
      - wallet
    patch:
      consumes:
      - application/json
      description: Update wallet
      parameters:
      - description: Body for update wallet
        in: body
        name: UpdateWallet
        required: true
        schema:
          $ref: '#/definitions/wallet.UpdateWallet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Update wallet
      tags:
      - wallet
    post:
      consumes:
      - application/json
      description: Create wallet
      parameters:
      - description: Body for create wallet
        in: body
        name: CreateWallet
        required: true
        schema:
          $ref: '#/definitions/wallet.CreateWallet'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/wallet.Wallet'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/wallet.Err'
      summary: Create wallet
      tags:
      - wallet
swagger: "2.0"